requirements.txt
pytest
playwright
pytest-playwright
load_dotenv